requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
PyMuPDF
langchain
langchain-google-genai
langchain-community 
faiss-cpu
ngrok
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pinecone
requests
pydantic-settings
beanie
motor